data/requirements.txt
paste==3.4.4
psycopg2==2.8.6
PyYaml==5.3.1
requests==2.24.0
webapp2>=3.0.0b1
webob==1.8.6
numpy==1.19.2
scipy==1.5.2
# smart-open==1.10.0  # solves bug in printing an warning message https://github.com/RaRe-Technologies/smart_open/issues/475
smart-open==2.2.0
gensim==3.8.3
